requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
chromadb==0.5.0
fastapi==0.111.0
langchain==0.2.4
langchain-community==0.2.4
langchain-core==0.2.6
langchain-openai==0.1.8
pydantic==2.7.3
python-dotenv==1.0.1
tenacity==8.3.0
azure-ai-textanalytics==5.3.0
pdfplumber==0.11.1 


annotated-types==0.7.0
anyio==4.4.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-search-documents==11.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
openai==1.35.4
packaging==24.1
preshed==3.0.9
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0